requirements.txt
streamlit
streamlit-image-comparison
streamlit-toggle-switch
streamlit-image-select
opencv-python
pillow
Image

